convert unsigned ints to ints

+++ src/game.cc
@@ -22,12 +22,12 @@ bool Game::operator==(const Game & game) const {
 }
 
 // Getter methods
-uint Game::white_moves() const { return m_white_moves; }
-uint Game::black_moves() const { return m_black_moves; }
+int Game::white_moves() const { return m_white_moves; }
+int Game::black_moves() const { return m_black_moves; }
 State Game::getState() const { return m_state; }
 Side Game::active_side() const { return m_active_side; }
 Side Game::winning_side() const {return m_winning_side;}
-uint Game::moves_left() const { return m_moves_left; }
+int Game::moves_left() const { return m_moves_left; }
 
 std::array<Location, 4> & Game::active_players() {
     return m_active_side == Side::WHITE? m_whiteLocations : m_blackLocations;
@@ -56,18 +56,18 @@ void Game::nextTurn() {
 
 void Game::setGameOver(Side winner) { m_state = State::ENDED; m_winning_side = winner; }
 
-constexpr uint Game::countInnerWalls(Location const start, Location const end) const {
+constexpr int Game::countInnerWalls(Location const start, Location const end) const {
 
     // differences
-    int const dx = static_cast<int>(end.x - start.x);
-    int const dy = static_cast<int>(end.y - start.y);
+	int const dx = end.x - start.x;
+	int const dy = end.y - start.y;
 
     // block coordinates
-    uint const sx = uint(start.x / 2);
-    uint const sy = uint(start.y / 2);
+	uint const sx = uint(start.x / 2);
+	uint const sy = uint(start.y / 2);
 
-    uint const ex = uint(end.x / 2);
-    uint const ey = uint(end.y / 2);
+	uint const ex = uint(end.x / 2);
+	uint const ey = uint(end.y / 2);
     WallConfig const & wstart = m_wall_matrix[sy][sx];
     WallConfig const & wend = m_wall_matrix[ey][ex];
 
@@ -76,9 +76,9 @@ constexpr uint Game::countInnerWalls(Location const start, Location const end) c
 
     int const first = (dx ? 1 : -1);
     int const second = (sx == ex && sy == ey ? 0 : dx + dy);
-    int const third = -1 + 2 * cabs(dx) * static_cast<int>(start.y % 2) + 2 * cabs(dy) * static_cast<int>(start.x % 2);
+	int const third = -1 + 2 * cabs(dx) * (start.y % 2) + 2 * cabs(dy) * (start.x % 2);
 
-    uint num = 0;
+	int num = 0;
     if (second) {
         num += wall::hasWallSeg(wend, {first, -second, third});
     }
@@ -98,14 +98,14 @@ constexpr bool Game::existsPlayerAtLocation(Location const location) const {
     return toReturn;
 }
 
-constexpr std::optional<uint> Game::moveCost(Location old_location, Direction direction) const {
+constexpr std::optional<int> Game::moveCost(Location old_location, Direction direction) const {
     if (std::optional<Location> new_location = old_location + direction) {
         return countInnerWalls(old_location, new_location.value());
     }
     return {};
 }
 
-constexpr std::optional<uint> Game::moveCost(Location oldL, Location newL) const {
+constexpr std::optional<int> Game::moveCost(Location oldL, Location newL) const {
 
     if (oldL == newL) return 0;
 
@@ -113,8 +113,8 @@ constexpr std::optional<uint> Game::moveCost(Location oldL, Location newL) const
 
     if (existsPlayerAtLocation(newL)) return {};
 
-    int wallCount = 0;
-    int distance = abs(static_cast<int>(newL.x - oldL.x)) + abs(static_cast<int>(newL.y - oldL.y));
+	int wallCount = 0;
+	int distance = abs(newL.x - oldL.x) + abs(newL.y - oldL.y);
     if (distance > 2) return {};
     if (distance == 2) {
         Location const connecting = {(newL.x + oldL.x) / 2, (newL.y + oldL.y) / 2};
@@ -130,27 +130,27 @@ constexpr std::optional<uint> Game::moveCost(Location oldL, Location newL) const
 }
 
 
-std::optional<uint> Game::getPlayerNumber(Location location) {
+std::optional<int> Game::getPlayerNumber(Location location) {
     for (size_t i = 0; i != 4; ++i) {
         if( active_players()[i] == location ) return i;
     }
     return {};
 }
 
-Location Game::getPlayerLocation(uint player_num) {
-    return active_players()[player_num];
+Location Game::getPlayerLocation(int player_num) {
+	return active_players()[static_cast<uint>(player_num)];
 }
 
-bool Game::decrementMoves(uint moves) {
+bool Game::decrementMoves(int moves) {
     if (m_moves_left < moves ) return false;
     (m_active_side == Side::WHITE ? m_white_moves : m_black_moves) += moves;
     m_moves_left -= moves;
     return true;
 }
 
-std::optional<Location> Game::movePlayer(uint player_num, Direction direction) {
+std::optional<Location> Game::movePlayer(int player_num, Direction direction) {
     Location old_location = getPlayerLocation(player_num);
-    std::optional<uint> cost;
+	std::optional<int> cost;
     std::optional<Location> new_location = old_location + direction;
 
     if (new_location){
@@ -176,7 +176,7 @@ std::optional<Location> Game::movePlayer(uint player_num, Direction direction) {
 std::optional<Location> Game::movePlayer(Location old_location, Location new_location){
     auto player = getPlayerNumber(old_location);
     if (!player) return {};
-    std::optional<uint> cost = moveCost(old_location, new_location);
+	std::optional<int> cost = moveCost(old_location, new_location);
 
     if (cost && decrementMoves(cost.value())){
         // the moves have been 'paid' successfully, we now have to move to player
@@ -190,7 +190,7 @@ std::optional<Location> Game::movePlayer(Location old_location, Location new_loc
     return new_location;
 }
 
-void Game::setPlayerLocation(uint player, Location location) {
-    active_players()[player] = location;
+void Game::setPlayerLocation(int player, Location location) {
+	active_players()[static_cast<uint>(player)] = location;
 }
